Refactor package name to tilesMetaDataOnboarding
