+++ src/main/java/ogc/rs/processes/ProcessesRunnerImpl.java
@@ -15,7 +15,7 @@ import io.vertx.sqlclient.Tuple;
 import ogc.rs.common.DataFromS3;
 import ogc.rs.processes.collectionAppending.CollectionAppendingProcess;
 import ogc.rs.processes.collectionOnboarding.CollectionOnboardingProcess;
-import ogc.rs.processes.tilesOnboarding.TilesMetaDataOnboardingProcess;
+import ogc.rs.processes.tilesMetaDataOnboarding.TilesMetaDataOnboardingProcess;
 import ogc.rs.processes.util.Status;
 import ogc.rs.processes.util.UtilClass;
 import org.apache.logging.log4j.LogManager;
